Fixed CoAP format to be a uint16_t since the enum might compile to 8 bit and cause problems if 16-bit format types are used

+++ Huawei_LiteOS/components/connectivity/lwm2m/core/er-coap-13/er-coap-13.h
@@ -277,13 +277,13 @@ typedef struct {
 /* Option format serialization*/
 #define COAP_SERIALIZE_INT_OPTION(number, field, text)  \
     if (IS_OPTION(coap_pkt, number)) { \
-      PRINTF(text" [%u]\n", coap_pkt->field); \
+      PRINTF(text" [%u]\n", (unsigned int)coap_pkt->field); \
       option += coap_serialize_int_option(number, current_number, option, coap_pkt->field); \
       current_number = number; \
     }
 #define COAP_SERIALIZE_BYTE_OPTION(number, field, text)      \
     if (IS_OPTION(coap_pkt, number)) { \
-      PRINTF(text" %u [0x%02X%02X%02X%02X%02X%02X%02X%02X]\n", coap_pkt->field##_len, \
+      PRINTF(text" %u [0x%02X%02X%02X%02X%02X%02X%02X%02X]\n", (unsigned int)coap_pkt->field##_len, \
         coap_pkt->field[0], \
         coap_pkt->field[1], \
         coap_pkt->field[2], \
@@ -298,7 +298,7 @@ typedef struct {
     }
 #define COAP_SERIALIZE_STRING_OPTION(number, field, splitter, text)      \
     if (IS_OPTION(coap_pkt, number)) { \
-      PRINTF(text" [%.*s]\n", coap_pkt->field##_len, coap_pkt->field); \
+      PRINTF(text" [%.*s]\n", (int)coap_pkt->field##_len, coap_pkt->field); \
       option += coap_serialize_array_option(number, current_number, option, (uint8_t *) coap_pkt->field, coap_pkt->field##_len, splitter); \
       current_number = number; \
     }
@@ -322,10 +322,10 @@ typedef struct {
     if (IS_OPTION(coap_pkt, number)) \
     { \
       uint32_t block = coap_pkt->field##_num << 4; \
-      PRINTF(text" [%lu%s (%u B/blk)]\n", coap_pkt->field##_num, coap_pkt->field##_more ? "+" : "", coap_pkt->field##_size); \
+      PRINTF(text" [%lu%s (%u B/blk)]\n", (unsigned long)coap_pkt->field##_num, coap_pkt->field##_more ? "+" : "", coap_pkt->field##_size); \
       if (coap_pkt->field##_more) block |= 0x8; \
       block |= 0xF & coap_log_2(coap_pkt->field##_size/16); \
-      PRINTF(text" encoded: 0x%lX\n", block); \
+      PRINTF(text" encoded: 0x%lX\n", (unsigned long)block); \
       option += coap_serialize_int_option(number, current_number, option, block); \
       current_number = number; \
     }
